+++ src/ggrc-client/js/components/custom-roles/tests/custom-roles_spec.js
@@ -3,15 +3,13 @@
   Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
 
-import Component from '../custom-roles';
-import {getComponentVM} from '../../../../js_specs/spec_helpers';
-import DeferredTransaction from '../../../plugins/utils/deferred-transaction-utils';
+import ViewModel from '../custom-roles-vm';
 
-describe('custom-roles component', () => {
+describe('custom-roles view model', () => {
   let vm;
 
   beforeEach(() => {
-    vm = getComponentVM(Component);
+    vm = new ViewModel();
   });
 
   describe('isReadonly prop', () => {
@@ -132,45 +130,6 @@ describe('custom-roles component', () => {
       it('sets null to updatableGroupId attribute', () => {
         vm.save(args);
         expect(vm.attr('updatableGroupId')).toBe(null);
-      });
-    });
-
-    describe('if deferredSave is defined', () => {
-      let instanceSave;
-
-      beforeEach(() => {
-        instanceSave = $.Deferred();
-        vm.attr('instance', {
-          save: () => instanceSave,
-        });
-        vm.attr('deferredSave', new DeferredTransaction((resolve, reject) => {
-          vm.attr('instance').save().done(resolve).fail(reject);
-        }, 0));
-        spyOn(vm, 'filterACL');
-      });
-
-      it('pushes callback into deferredSave which sets updatableGroupId',
-        (done) => {
-          let pushSpy = spyOn(vm.attr('deferredSave'), 'push')
-            .and.returnValue(instanceSave);
-          args.groupId = 711;
-
-          vm.save(args);
-          pushSpy.calls.allArgs()[0][0]();
-          expect(vm.attr('updatableGroupId')).toBe(args.groupId);
-          done();
-        });
-
-      it('calls filterACL after save', (done) => {
-        spyOn(vm.attr('deferredSave'), 'push')
-          .and.returnValue(instanceSave.resolve());
-
-        vm.save(args);
-
-        instanceSave.then(() => {
-          expect(vm.filterACL).toHaveBeenCalled();
-          done();
-        });
       });
     });
   });
